+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 0.11.9
+
+### Patch Changes
+
+- Small bugs
+- Updated dependencies
+  - @forklaunch/validator@0.10.32
+  - @forklaunch/common@0.6.32
+  - @forklaunch/core@0.18.5
+  - @forklaunch/ws@0.2.3
+
 ## 0.11.8
 
 ### Patch Changes