+++ framework/testing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/testing
 
+## 0.0.21
+
+### Patch Changes
+
+- Small bugs
+
 ## 0.0.20
 
 ### Patch Changes