+++ framework/ws/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @forklaunch/ws
 
+## 0.2.3
+
+### Patch Changes
+
+- Small bugs
+- Updated dependencies
+  - @forklaunch/validator@0.10.32
+  - @forklaunch/common@0.6.32
+  - @forklaunch/core@0.18.5
+
 ## 0.2.2
 
 ### Patch Changes